Refactoring StopsHandler to use shared functions
