+++ traffic_stops/static/js/app/common/Stops.js
@@ -1,4 +1,5 @@
 import _ from 'underscore';
+import d3 from 'd3';
 
 /***
  * StopsHandler data processing helper functions
@@ -27,4 +28,44 @@ export function build_totals (data) {
   }
 
   return total;
+}
+
+export function build_pie_data (data, total, Stops) {
+  let pie = d3.map();
+
+  data.forEach((v) => {
+    if (v.year >= Stops.start_year) pie.set(v.year, d3.map(v));
+  });
+
+  pie.set('Total', d3.map(total));
+
+  return pie;
+}
+
+export function get_total_by_type (dataType, yr) {
+  var total = 0;
+  dataType.forEach((type) => {
+    total += (typeof yr[type] === 'undefined') ? 0 : yr[type];
+  });
+  return total;
+}
+
+export function build_line_data (data, types, Stops) {
+  var line = d3.map();
+
+  types.forEach((dataType) => {
+    dataType.forEach((v) => {
+      line.set(v, []);
+    });
+    data.forEach((yr) => {
+      if (yr.year >= Stops.start_year) {
+        var total = get_total_by_type(dataType, yr);
+        dataType.forEach((type) => {
+          line.get(type).push({x: yr.year, y:(yr[type] > 0 ? yr[type]/total : 0)});
+        })
+      }
+    })
+  });
+
+  return line;
 }
