+++ traffic_stops/static/js/app/states/md/Stops.js
@@ -3,7 +3,7 @@ import VisualBase from '../../base/VisualBase.js';
 import TableBase from '../../base/TableBase.js';
 import Stops from './defaults.js';
 
-import { build_totals } from '../common/Stops.js';
+import * as C from '../../common/Stops.js';
 
 import Backbone from 'backbone';
 import _ from 'underscore';
@@ -11,55 +11,18 @@ import d3 from 'd3';
 import $ from 'jquery';
 Backbone.$ = $;
 
-export function build_pie_data (data, total, Stops) {
-  let pie = d3.map();
-
-  data.forEach((v) => {
-    if (v.year >= Stops.start_year) pie.set(v.year, d3.map(v));
-  });
-
-  pie.set('Total', d3.map(total));
-
-  return pie;
-}
-
-export function get_total_by_race (dataType, yr) {
-  var total = 0;
-  dataType.forEach((ethnicity) => {
-    total += (typeof yr[ethnicity] === 'undefined') ? 0 : yr[ethnicity];
-  });
-  return total;
-}
-
-export function build_line_data (data, Stops) {
-  var line = d3.map();
-
-  Stops.ethnicities.forEach((v) => {
-    line.set(v, []);
-  });
-  data.forEach((yr) => {
-    if (yr.year >= Stops.start_year) {
-      var total = get_total_by_race(Stops.ethnicities, yr);
-      Stops.ethnicities.forEach((ethnicity) => {
-        line.get(ethnicity).push({x: yr.year, y:(yr[ethnicity] > 0 ? yr[ethnicity]/total : 0)});
-      })
-    }
-  })
-  return line;
-}
-
 export var StopsHandler = DataHandlerBase.extend({
   clean_data: function () {
 
     // build totals
     var data = this.get('raw_data');
-    var total = build_totals(data);
+    var total = C.build_totals(data);
 
-    // build-data for pie-chart
-    var pie = build_pie_data(data, total, Stops);
+    // build data for pie-chart
+    var pie = C.build_pie_data(data, total, Stops);
 
     // build data for line-chart
-    var line = build_line_data(data, Stops);
+    var line = C.build_line_data(data, [Stops.ethnicities], Stops);
 
     // set object data
     this.set('data', {
@@ -71,65 +34,12 @@ export var StopsHandler = DataHandlerBase.extend({
   }
 });
 
-export var StopRatioDonut = VisualBase.extend({
+export var StopRatioDonut = C.StopRatioDonutBase.extend({
   defaults: {
     width: 300,
     height: 300
   },
 
-  setDefaultChart: function(){
-    this.chart = nv.models.pieChart()
-      .x((d) => d.key)
-      .y((d) => d.value)
-      .color((d) => d.data.color)
-      .width(this.get("width"))
-      .height(this.get("height"))
-      .showLabels(true)
-      .labelType("percent")
-      .donutRatio(0.35)
-      .labelThreshold(0.05)
-      .donut(true)
-      .tooltipContent((key, y, e, graph) => (
-        `<h3 class="stops donut-label">${ key }</h3><p>${ y.replace(/\.\d*/, '') }</p>`
-      ));
-  },
-  drawStartup: function(){
-
-    // get year options for pulldown menu
-    var selector = $('<select>'),
-        year_options = this.data.pie.keys(),
-        opts = year_options.map((v) => `<option value="${v}">${v}</option>`),
-        getData = () => {
-          var value = selector.val();
-          this.dataset =  this.data.pie.get(value);
-          this.drawChart();
-        };
-
-    selector
-      .append(opts)
-      .val("Total")
-      .on('change', getData);
-
-    $('<div>')
-      .html(selector)
-      .appendTo(this.div);
-
-    getData();
-  },
-  drawChart: function(){
-    var data = this._formatData();
-
-    nv.addGraph(() => {
-      d3.select(this.svg[0])
-          .datum(data)
-        .transition().duration(1200)
-          .attr('width', "100%")
-          .attr('height', "100%")
-          .attr("preserveAspectRatio", "xMinYMin")
-          .attr('viewBox', `0 0 ${this.get('width')} ${this.get('height')}`)
-          .call(this.chart);
-    });
-  },
   _formatData: function(){
     var data = [],
         selected = this.dataset,
@@ -147,6 +57,7 @@ export var StopRatioDonut = VisualBase.extend({
 
     return data;
   },
+  
   triggerRaceToggle: () => null
 });
